+++ src/components/Header/Header.tsx
@@ -1,23 +1,10 @@
 import React from 'react';
-import { useDispatch, useSelector } from "react-redux";
-import { change } from "../../features/language/languageSlice";
 import { LanguageDropDown } from 'components';
 
 interface HeaderProps {
-  category: string;
   pageTitle: string;
 };
-
-interface RootState {
-  language: {
-    value: string;
-  }  
-}
-
-const Header: React.FC<HeaderProps> = ({ pageTitle, category }) => {
-
-  const dispatch = useDispatch();
-  const language = useSelector((state: RootState) => state.language.value);
+const Header: React.FC<HeaderProps> = ({ pageTitle }) => {
 
   return (
     <div
